Move the Android Studio Plugin packages to the NAPPA namespace under the S2 Group namespace
Issue: S2-group/NAPPA#29
