+++ Android-Studio-Plugin/src/main/java/nl/vu/cs/s2group/nappa/plugin/action/TestAction.java
@@ -1,4 +1,4 @@
-package test;
+package nl.vu.cs.s2group.nappa.plugin.action;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
@@ -245,7 +245,7 @@ public class TestAction extends AnAction {
                      *      the prefetching lib's package
                      * */
                     final PsiElement importToAdd = PsiElementFactory.SERVICE.getInstance(project).createImportStatementOnDemand(
-                            "nl.vu.cs.s2group");
+                            "nl.vu.cs.s2group.nappa");
 
                     // Fetch all the import statements for this java file
                     PsiImportList importList = javaFile.getImportList();
@@ -256,7 +256,7 @@ public class TestAction extends AnAction {
                     // Check to see that the prefetch lib import has already taken place, to
                     //      avoid duplicates
                     for (PsiImportStatement importStatement : importStatements) {
-                        if (importStatement.getText().compareTo("import nl.vu.cs.s2group.*;") == 0) {
+                        if (importStatement.getText().compareTo("import nl.vu.cs.s2group.nappa*;") == 0) {
                             imported = true;
                             break;
                         }
